requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
docx==0.2.4
email_validator==2.2.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.3.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
lxml==5.3.1
MarkupSafe==3.0.2
openai==1.65.1
packaging==24.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-docx==1.1.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.38
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3
